requirements.txt
absl-py==1.4.0
asgiref==3.4.1
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.3.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-cors-headers==3.13.0
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
faiss-cpu==1.7.2
filelock==3.4.2
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
huggingface-hub==0.12.0
idna==3.3
joblib==1.1.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.2
nltk==3.6.7
numpy==1.22.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.0.0
tokenizers==0.11.4
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
transformers==4.26.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.2.3
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
